third_party/WebKit/LayoutTests/fast/forms/date-multiple-fields/date-clearbutton-preventdefault-mousecapture-status.html

<!DOCTYPE>
<html>
<head>
<script src="../resources/common.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body>

<input id="input" type="date" style="font-family:ahem; font-size:16px;">
<button id="js-btn-test" type="button"></button>

<script>
description('Tests, after clicking mouse on clearInputButton of date input type, will release mouse capture in case of preventdefault.');

var input = document.getElementById('input');
input.value = '2345-07-19';

var btn = document.getElementById('js-btn-test');
var output = '';

btn.addEventListener('click', function ()
{
    output = 'Clicked';
});

document.addEventListener('mouseup', function (evt)
{
    evt.preventDefault();
});

var clearButton = getElementByPseudoId(internals.shadowRoot(input),'-webkit-clear-button');
clickElement(clearButton);
shouldBeEqualToString('input.value', '');

clickElement(btn);
shouldBeEqualToString('output', 'Clicked');
</script>
</body>
</html>
