third_party/WebKit/LayoutTests/fast/forms/date-multiple-fields/date-clearbutton-preventdefault-mousecapture-status-expected.txt

Tests, after clicking mouse on clearInputButton of date input type, will release mouse capture in case of preventdefault.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is ""
PASS output is "Clicked"
PASS successfullyParsed is true

TEST COMPLETE
 
